Split MusicXML into import/export/shared folders
