+++ src/importexport/musicxml/internal/musicxml/musicxml.cmake
@@ -1,32 +1,32 @@
 
 set (MUSICXML_SRC
-    ${CMAKE_CURRENT_LIST_DIR}/exportmusicxml.cpp
-    ${CMAKE_CURRENT_LIST_DIR}/exportmusicxml.h
-    ${CMAKE_CURRENT_LIST_DIR}/importmusicxml.cpp
-    ${CMAKE_CURRENT_LIST_DIR}/importmusicxml.h
-    ${CMAKE_CURRENT_LIST_DIR}/importmusicxmllogger.cpp
-    ${CMAKE_CURRENT_LIST_DIR}/importmusicxmllogger.h
-    ${CMAKE_CURRENT_LIST_DIR}/importmusicxmlnoteduration.cpp
-    ${CMAKE_CURRENT_LIST_DIR}/importmusicxmlnoteduration.h
-    ${CMAKE_CURRENT_LIST_DIR}/importmusicxmlnotepitch.cpp
-    ${CMAKE_CURRENT_LIST_DIR}/importmusicxmlnotepitch.h
-    ${CMAKE_CURRENT_LIST_DIR}/importmusicxmlpass1.cpp
-    ${CMAKE_CURRENT_LIST_DIR}/importmusicxmlpass1.h
-    ${CMAKE_CURRENT_LIST_DIR}/importmusicxmlpass2.cpp
-    ${CMAKE_CURRENT_LIST_DIR}/importmusicxmlpass2.h
-    ${CMAKE_CURRENT_LIST_DIR}/importxml.cpp
-    ${CMAKE_CURRENT_LIST_DIR}/musicxmlfonthandler.cpp
-    ${CMAKE_CURRENT_LIST_DIR}/musicxmlfonthandler.h
-    ${CMAKE_CURRENT_LIST_DIR}/musicxmlpart.cpp
-    ${CMAKE_CURRENT_LIST_DIR}/musicxmlpart.h
-    ${CMAKE_CURRENT_LIST_DIR}/musicxmlsupport.cpp
-    ${CMAKE_CURRENT_LIST_DIR}/musicxmlsupport.h
-    ${CMAKE_CURRENT_LIST_DIR}/musicxmltupletstate.cpp
-    ${CMAKE_CURRENT_LIST_DIR}/musicxmltupletstate.h
-    ${CMAKE_CURRENT_LIST_DIR}/musicxmltypes.cpp
-    ${CMAKE_CURRENT_LIST_DIR}/musicxmltypes.h
-    ${CMAKE_CURRENT_LIST_DIR}/musicxmlvalidation.cpp
-    ${CMAKE_CURRENT_LIST_DIR}/musicxmlvalidation.h
-    ${CMAKE_CURRENT_LIST_DIR}/musicxmlvoicedesc.cpp
-    ${CMAKE_CURRENT_LIST_DIR}/musicxmlvoicedesc.h
+    ${CMAKE_CURRENT_LIST_DIR}/export/exportmusicxml.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/export/exportmusicxml.h
+    ${CMAKE_CURRENT_LIST_DIR}/import/importmusicxml.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/import/importmusicxml.h
+    ${CMAKE_CURRENT_LIST_DIR}/import/importmusicxmllogger.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/import/importmusicxmllogger.h
+    ${CMAKE_CURRENT_LIST_DIR}/import/importmusicxmlnoteduration.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/import/importmusicxmlnoteduration.h
+    ${CMAKE_CURRENT_LIST_DIR}/import/importmusicxmlnotepitch.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/import/importmusicxmlnotepitch.h
+    ${CMAKE_CURRENT_LIST_DIR}/import/importmusicxmlpass1.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/import/importmusicxmlpass1.h
+    ${CMAKE_CURRENT_LIST_DIR}/import/importmusicxmlpass2.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/import/importmusicxmlpass2.h
+    ${CMAKE_CURRENT_LIST_DIR}/import/importxml.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/import/musicxmlpart.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/import/musicxmlpart.h
+    ${CMAKE_CURRENT_LIST_DIR}/import/musicxmltupletstate.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/import/musicxmltupletstate.h
+    ${CMAKE_CURRENT_LIST_DIR}/import/musicxmlvalidation.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/import/musicxmlvalidation.h
+    ${CMAKE_CURRENT_LIST_DIR}/import/musicxmlvoicedesc.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/import/musicxmlvoicedesc.h
+    ${CMAKE_CURRENT_LIST_DIR}/shared/musicxmlfonthandler.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/shared/musicxmlfonthandler.h
+    ${CMAKE_CURRENT_LIST_DIR}/shared/musicxmlsupport.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/shared/musicxmlsupport.h
+    ${CMAKE_CURRENT_LIST_DIR}/shared/musicxmltypes.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/shared/musicxmltypes.h
     )
