+++ src/importexport/musicxml/internal/musicxml/shared/musicxmltypes.h
