+++ src/importexport/musicxml/internal/mxlwriter.cpp
@@ -22,7 +22,7 @@
 
 #include "mxlwriter.h"
 
-#include "musicxml/exportmusicxml.h"
+#include "musicxml/export/exportmusicxml.h"
 
 #include "log.h"
